+++ doc/docs/tutorials/getting-started/bin/main.ml
@@ -1,6 +1,3 @@
 let () =
-  Commandlang_to_cmdliner.run
-    Getting_started.cmd
-    ~name:"my-calculator"
-    ~version:"%%VERSION%%"
+  Cmdlang_to_cmdliner.run Getting_started.cmd ~name:"my-calculator" ~version:"%%VERSION%%"
 ;;