+++ lib/cmdlang_to_climate/src/translate.mli
@@ -1,0 +1,1 @@
+val command : 'a Cmdlang.Command.t -> 'a Climate.Command.t