+++ lib/commandlang_to_climate/src/translate.mli
@@ -1,1 +1,0 @@
-val command : 'a Commandlang.Command.t -> 'a Climate.Command.t