+++ lib/err/test/cram/main.ml
@@ -43,4 +43,4 @@ let write_cmd =
 ;;
 
 let main = Command.group ~summary:"test err from the command line" [ "write", write_cmd ]
-let () = Commandlang_to_cmdliner.run main ~name:"main" ~version:"%%VERSION%%"
+let () = Cmdlang_to_cmdliner.run main ~name:"main" ~version:"%%VERSION%%"