+++ test/base/bin/main_base.ml
@@ -1,1 +1,1 @@
-let () = Command_unix.run (Commandlang_to_base.Translate.unit Test_command.cmd)
+let () = Command_unix.run (Cmdlang_to_base.Translate.unit Test_command.cmd)