+++ test/climate/bin/main_climate.ml
@@ -1,1 +1,1 @@
-let () = Climate.Command.run (Commandlang_to_climate.Translate.command Test_command.cmd)
+let () = Climate.Command.run (Cmdlang_to_climate.Translate.command Test_command.cmd)