+++ test/cmdliner/bin/main_cmdliner.ml
@@ -1,3 +1,3 @@
 let () =
-  Commandlang_to_cmdliner.run Test_command.cmd ~name:Sys.argv.(0) ~version:"%%VERSION%%"
+  Cmdlang_to_cmdliner.run Test_command.cmd ~name:Sys.argv.(0) ~version:"%%VERSION%%"
 ;;